+++ piwigo/Image/Parameters/Cells/EditImageThumbTableViewCell.swift
@@ -84,7 +84,7 @@ extension EditImageThumbTableViewCell: UICollectionViewDelegateFlowLayout
     }
 
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        return CGFloat(AlbumUtilities.kImageDetailsCellSpacing)
+        return AlbumUtilities.kImageDetailsCellSpacing
     }
 
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
